+++ R/checks.R
@@ -40,7 +40,7 @@
       },
       error = function(cond) {
         message(sprintf("'%s' is not convertible to a terra SpatRaster object!", name))
-        message(sprintf("'%s' must be a SpatRaster, stars, Raster* object, or path to raster a file on disk.", name))
+        message(sprintf("'%s' must be a SpatRaster, stars, Raster* object, or path to a raster file on disk.", name))
       }
     )
   }